requirements.txt
tensorflow==2.21.0
tensorflowjs==4.21.0
setuptools<81
fastapi==0.115.13
uvicorn==0.34.3
httpx==0.28.1
numpy==2.1.3
faiss-cpu==1.11.0
sentence-transformers==4.1.0
python-dotenv==1.1.1
